ptc: composition (Children props)
