+++ react-guide/src/components/Card.js
@@ -1,0 +1,7 @@
+import React from "react";
+import "./Card.css";
+
+export default function Card(props) {
+  const classes = "card" + props.className;
+  return <div className={classes}>{props.children}</div>;
+}
