+++ react-guide/src/components/ExpenseItem.js
@@ -1,17 +1,14 @@
 import React from "react";
+import Card from "./Card";
 import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
 
 export default function ExpenseItem(props) {
   return (
-    <div>
-      <div className="expense-item">
-        <ExpenseDate date={props.date} />
-      </div>
-      <div className="expense-item__description">
-        <h2>{props.title} </h2>
-      </div>
+    <Card className="expense-item">
+      <ExpenseDate date={props.date} />
+      <div className="expense-item__description">{props.title}</div>
       <div className="expense-item__price">{props.amount}</div>
-    </div>
+    </Card>
   );
 }
